+++ api/sanPham/delete.php
@@ -6,37 +6,46 @@
 
     include_once('../../config/db.php');
     include_once('../../model/sanPham.php');
+    include_once '../../middleware/check_role.php';
 
-    // Kiểm tra nếu phương thức là DELETE
-    if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
-        // Kết nối tới cơ sở dữ liệu
-        $db = new db();
-        $connect = $db->connect();
 
-        $sanPham = new SanPham($connect);
-
-        // Kiểm tra nếu có ID được cung cấp
-        if (isset($_GET['id'])) {
-            $sanPham->product_id = $_GET['id'];
-
-            // Kiểm tra nếu sản phẩm tồn tại
-            if ($sanPham->exists()) {
-                // Nếu tồn tại, tiến hành xoá
-                if ($sanPham->delete()) {
-                    echo json_encode(array('message' => 'sanPham Deleted'));
+    if(checkRole('admin')) {
+        // Kiểm tra nếu phương thức là DELETE
+        if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
+            // Kết nối tới cơ sở dữ liệu
+            $db = new db();
+            $connect = $db->connect();
+    
+            $sanPham = new SanPham($connect);
+    
+            // Kiểm tra nếu có ID được cung cấp
+            if (isset($_GET['id'])) {
+                $sanPham->product_id = $_GET['id'];
+    
+                // Kiểm tra nếu sản phẩm tồn tại
+                if ($sanPham->exists()) {
+                    // Nếu tồn tại, tiến hành xoá
+                    if ($sanPham->delete()) {
+                        echo json_encode(array('message' => 'sanPham Deleted'));
+                    } else {
+                        echo json_encode(array('message' => 'sanPham Not Deleted'));
+                    }
                 } else {
-                    echo json_encode(array('message' => 'sanPham Not Deleted'));
+                    // Nếu không tồn tại, trả về thông báo
+                    echo json_encode(array('message' => 'sanPham ID does not exist'));
                 }
             } else {
-                // Nếu không tồn tại, trả về thông báo
-                echo json_encode(array('message' => 'sanPham ID does not exist'));
+                // Nếu không có ID trên URL, trả về thông báo
+                echo json_encode(array('message' => 'No sanPham ID provided'));
             }
         } else {
-            // Nếu không có ID trên URL, trả về thông báo
-            echo json_encode(array('message' => 'No sanPham ID provided'));
+            // Nếu không phải là phương thức DELETE, trả về thông báo lỗi
+            echo json_encode(array('message' => 'Invalid request method. Only DELETE requests are allowed.'));
         }
-    } else {
-        // Nếu không phải là phương thức DELETE, trả về thông báo lỗi
-        echo json_encode(array('message' => 'Invalid request method. Only DELETE requests are allowed.'));
     }
+    else {
+        exit();
+    }
+
+    
 ?>
